scrapers/bluesky/requirements.txt
fastapi
uvicorn[standard]
httpx
groq
pydantic>=2
python-dotenv
